+++ crates/sommelier-auction/src/bid.rs
@@ -5,6 +5,6 @@ use crate::denom::Denom;
 pub struct Bid {
     pub auction_id: u32,
     pub fee_token: Denom,
-    pub maximum_usomm_in: u64,
-    pub minimum_tokens_out: u64,
+    pub maximum_usomm_in: u128,
+    pub minimum_tokens_out: u128,
 }
